Add Taggable concern for models
issue #18
